bank plugin: Remove ContainerCalculation class
This commit removes the hashing feature from container calculation which
alone would make obsolete the ContainerCalculation class, so it
additionally inlines the `calculate()` method into the plugin class.

Additionally, this commit adds null handling for the items parameter of
`calculate()` to prevent NPEs.
