--- runelite-client/src/main/java/net/runelite/client/plugins/bank/ContainerCalculation.java
@@ -1,111 +1,0 @@
-/*
- * Copyright (c) 2019, TheStonedTurtle <https://github.com/TheStonedTurtle>
- * Copyright (c) 2019, Adam <Adam@sigterm.info>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.bank;
-
-import java.util.HashMap;
-import java.util.Map;
-import javax.annotation.Nullable;
-import javax.inject.Inject;
-import net.runelite.api.Constants;
-import net.runelite.api.Item;
-import net.runelite.api.ItemID;
-import net.runelite.client.game.ItemManager;
-
-class ContainerCalculation
-{
-	private final ItemManager itemManager;
-
-	private int hash;
-	private ContainerPrices containerPrices;
-
-	@Inject
-	private ContainerCalculation(ItemManager itemManager)
-	{
-		this.itemManager = itemManager;
-	}
-
-	@Nullable
-	ContainerPrices calculate(Item[] items)
-	{
-		// Returns last calculation if inventory hasn't changed
-		final int newHash =  hashItems(items);
-		if (containerPrices != null && hash == newHash)
-		{
-			return containerPrices;
-		}
-
-		hash = newHash;
-
-		long ge = 0;
-		long alch = 0;
-
-		for (final Item item : items)
-		{
-			final int qty = item.getQuantity();
-			final int id = item.getId();
-
-			if (id <= 0 || qty == 0)
-			{
-				continue;
-			}
-
-			switch (id)
-			{
-				case ItemID.COINS_995:
-					ge += qty;
-					alch += qty;
-					break;
-				case ItemID.PLATINUM_TOKEN:
-					ge += qty * 1000L;
-					alch += qty * 1000L;
-					break;
-				default:
-					final long storePrice = itemManager.getItemComposition(id).getPrice();
-					final long alchPrice = (long) (storePrice * Constants.HIGH_ALCHEMY_MULTIPLIER);
-					alch += alchPrice * qty;
-					ge += (long) itemManager.getItemPrice(id) * qty;
-					break;
-			}
-
-		}
-
-		ContainerPrices prices = new ContainerPrices(ge, alch);
-		containerPrices = prices;
-
-		return prices;
-	}
-
-	private int hashItems(final Item[] items)
-	{
-		final Map<Integer, Integer> mapCheck = new HashMap<>(items.length);
-		for (Item item : items)
-		{
-			mapCheck.put(item.getId(), item.getQuantity());
-		}
-
-		return mapCheck.hashCode();
-	}
-}
