+++ runelite-client/src/test/java/net/runelite/client/plugins/bank/BankPluginTest.java
@@ -25,6 +25,7 @@
  */
 package net.runelite.client.plugins.bank;
 
+import com.google.common.collect.ImmutableList;
 import com.google.inject.Guice;
 import com.google.inject.testing.fieldbinder.Bind;
 import com.google.inject.testing.fieldbinder.BoundFieldModule;
@@ -37,6 +38,7 @@ import net.runelite.api.ItemContainer;
 import net.runelite.api.ItemID;
 import net.runelite.client.game.ItemManager;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
@@ -104,5 +106,32 @@ public class BankPluginTest
 		assertFalse(bankPlugin.valueSearch(itemId, "ge > 0.02b"));
 
 		assertFalse(bankPlugin.valueSearch(itemId, "1000k"));
+	}
+
+	@Test
+	public void testCalculate()
+	{
+		Item coins = new Item(ItemID.COINS_995, Integer.MAX_VALUE);
+
+		Item whip = new Item(ItemID.ABYSSAL_WHIP, 1_000_000_000);
+
+		Item[] items = ImmutableList.of(
+			coins,
+			whip
+		).toArray(new Item[0]);
+
+		ItemComposition whipComp = mock(ItemComposition.class);
+		when(whipComp.getPrice())
+			.thenReturn(7); // 7 * .6 = 4, 4 * 1m overflows
+		when(itemManager.getItemComposition(ItemID.ABYSSAL_WHIP))
+			.thenReturn(whipComp);
+		when(itemManager.getItemPrice(ItemID.ABYSSAL_WHIP))
+			.thenReturn(3); // 1b * 3 overflows
+
+		final ContainerPrices prices = bankPlugin.calculate(items);
+		assertNotNull(prices);
+
+		assertTrue(prices.getHighAlchPrice() > Integer.MAX_VALUE);
+		assertTrue(prices.getGePrice() > Integer.MAX_VALUE);
 	}
 }
